modified Payout and Remittance Charge Commission in new format.

+++ setting-service/src/main/java/com/toshaniFintech/user_service/dto/response/RemittanceChargeCommissionResponse.java
@@ -8,9 +8,9 @@ import java.math.BigDecimal;
 public class RemittanceChargeCommissionResponse {
     private String id;
     private Integer slNo;
-    private BigDecimal minAmount;
-    private BigDecimal maxAmount;
+    private Double minAmount;
+    private Double maxAmount;
     private String chargeType;
-    private BigDecimal charge;
+    private Double charge;
     private Boolean active;
 }
